fix(chat): der Schnitt lag im Schreiber, nicht im Widget
Die abgeschnittene lila Henry-Karte im Hauptchat war kein eigenes
Render-Widget. Das UI ist bereits EINE Pipeline: chat.tsx toStep() macht aus
jeder Nachricht denselben TStep, card_transcript.tsx zeichnet jede fremde
Textnachricht mit derselben CollapsibleMarkdown-Komponente; der lila Rahmen ist
nur byKind === "henry" (accent2), das Sternchen nur senderIcon. Der Client
faltet erst ab 1600 Zeichen, mit "mehr anzeigen".

Gekuerzt wurde beim SCHREIBEN. notice.short() (240 Zeichen / 2 Saetze, haengt
" …" an) wurde am 2026-08-30 in die zwei Chat-Schreiber verdrahtet. Beleg aus
daemon/copilot_log.json: der 14:08-Eintrag steht mit 241 Zeichen und " …" so
auf der Platte - abgeschnitten, bevor das Telefon ihn je gesehen hat. Die
cls:"bot"-Nachricht um 14:10 daneben hat 293 Zeichen und kein Ellipsis; genau
die Asymmetrie, die der Owner fotografiert hat.

Das Zwei-Satz-Gesetz gilt fuer Kanaele, die NICHT scrollen koennen - Push,
Uhr, Button-Header. Der Chat ist ein Transkript. Also:

- henry_broker._notify_owner: Chat bekommt den vollen Text (der Push behaelt
  sein eigenes 230er Budget). Genau das, was die Docstring der Funktion seit
  2026-08-28 vorschreibt und was die Zeile darunter verletzt hat.
- pm_comm._say: kein Clip mehr. Die Kuerze wird in den AUTOREN erzwungen und
  von test_notice_routing.py als short(line) == line geprueft.
- pm_comm._escalate: voller Text in den Chat, _short nur noch fuer den Push -
  und auf Wortgrenze statt rohem [:180].
- pm_comm._ask_owner: Henry bekommt die Frage GANZ statt der Button-Kuerzung;
  eine vorgekuerzte Frage uebernehmen und darueber berichten war die zweite
  Kuerzung derselben Nachricht.
- pm._needs_from_owner: rohes qs[0][:180] raus - ein Schnitt mitten im Wort.
- notice.short(): `chars` ist jetzt eine ECHTE Obergrenze, das " …" wird aus
  dem Budget bezahlt statt obendrauf gelegt (war chars+2, im Widerspruch zur
  eigenen Docstring).

Warum es zweimal als behoben galt: ops/tests/test_henry_chat_full_text.py
prueft seit dem 2026-08-28-Fix genau diese Eigenschaft - und war auf HEAD ROT
(2 FAIL). Der Gate ist per Dekret LIGHT und faehrt keine Unit-Suite, also hat
das niemand gesehen. Der Test ist wieder gruen und deckt jetzt auch die zweite
Tuer in dieselbe Karte ab (pm_comm._say/_escalate) sowie die Wortgrenze des
Pushs: 9 der 16 Pins fallen gegen den alten Code, alle 16 halten gegen den
neuen.

Gate gruen; test_notice_routing, test_cost_watch, test_chat_mirror,
test_escalations, test_henry_audit_context, test_henry_privilege_gate,
test_load_admission gruen. (test_pm_resilience "owner told the card is
unstuck" war schon auf HEAD rot - unberuehrt, gehoert nicht hierher.)

Reiner Daemon-Fix, kein App-Code: kein OTA noetig.

+++ cells/copilot/henry_broker.py
@@ -579,18 +579,23 @@ def _notify_owner(text, t):
     # broker's decisions. copilot.say is the same line the lane pipeline's
     # _say_card uses; best-effort like the push.
     #
-    # Clipped to the same two-sentence ceiling as every other automatic notice
-    # (owner decree 2026-08-30). Henry is an LLM writing free prose into the one
-    # chat the owner actually reads, so he is the likeliest future source of the
-    # wall the decree bans - the cap belongs at this edge, not in his mandate,
-    # where it would be a request rather than a rule. Only the CHAT copy is cut:
-    # the push has always had its own 230-char budget above, and the card audit
-    # below still gets the message WHOLE, so nothing is lost, it just stops
-    # being shouted (the same split _notify_owner's docstring already draws).
+    # WHOLE, like every other chat message. A notice.short() clip stood here for
+    # one day (2026-08-30) and the owner photographed the result the same
+    # afternoon: his 14:08 chat card ended "... kein Agent-Turn) …" mid-thought,
+    # while the ordinary bot bubbles right above it were intact. The reason it
+    # looked like a rogue widget is that it WAS a rogue clip - but on the write
+    # side, not the render side (see the module note above _decide's report-back).
+    #
+    # The two-sentence law is a law about CHANNELS THAT CANNOT SCROLL - the FCM
+    # push (its own 230 above) and a question's button header. The chat is a
+    # transcript: card_transcript.tsx already folds anything past 1600 chars
+    # behind a "mehr anzeigen" toggle, which gives the decree what it actually
+    # wanted (no wall) without amputating the sentence that carries Henry's
+    # point. Clipping here just moved the push's budget onto the transcript -
+    # precisely what this function's own docstring forbids.
     try:
         from cells.copilot import copilot
-        from spine.comms.notice import short
-        copilot.say(short(text), cls="pm", card=(t or {}).get("id") or None)
+        copilot.say(text, cls="pm", card=(t or {}).get("id") or None)
     except Exception:
         pass
     if t:
